requirements.txt
asgiref==3.9.1
Django==5.2.6
django-js-asset==3.1.2
django-mdeditor==0.1.20
django-mptt==0.15.0
Markdown==3.5.2
Pillow==10.0.0
sqlparse==0.5.3
typing_extensions==4.15.0
